Feat: remove course info entity
