+++ src/course-feedback/course-feedback.entity.ts
@@ -6,9 +6,7 @@ import {
   UpdateDateColumn,
 } from 'typeorm';
 import { PastExam } from './past-exam.entity';
-import { Review } from './review.entity';
-import { Comment } from './comment.entity';
-import { CourseInfo } from './course-info.entity';
+import { Comment, Review } from './content.entity';
 
 @Entity('CourseFeedbacks')
 export class CourseFeedback {
@@ -25,11 +23,6 @@ export class CourseFeedback {
 
   @OneToMany(() => PastExam, (pastExam) => pastExam.courseFeedback)
   pastExams: PastExam[];
-
-  @OneToMany(() => CourseInfo, (courseInfo) => courseInfo.courseFeedback, {
-    eager: true,
-  })
-  courseInfos: CourseInfo[];
 
   @CreateDateColumn()
   createdAt: Date;
