+++ src/course-feedback/course-feedback.module.ts
@@ -2,10 +2,8 @@ import { Module } from '@nestjs/common';
 import { CourseFeedbackService } from './course-feedback.service';
 import { CourseFeedbackController } from './course-feedback.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CourseInfo } from './course-info.entity';
 import { CourseModule } from '../course/course.module';
-import { Review } from './review.entity';
-import { Comment } from './comment.entity';
+import { Comment, Review } from './content.entity';
 import { PastExam } from './past-exam.entity';
 import { CourseFeedback } from './course-feedback.entity';
 import { Reaction } from './reaction.entity';
@@ -13,7 +11,6 @@ import { Reaction } from './reaction.entity';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
-      CourseInfo,
       CourseFeedback,
       Comment,
       Review,
